raifbot/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Peter Vajdecka <petervajdecka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
fastapi = "^0.103.1"
python-dotenv = "^0.21.0"
pinecone-client = "^2.0.0"
openai = "^1.10.0"
langchain = "^0.1.0"
pymongo = "^4.1.1"
gunicorn = "^21.2.0"
uvicorn = "^0.29.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
motor = "^3.4.0"
duckduckgo-search = "^3.9.4"
tiktoken = "^0.6.0"
langchain-community = "^0.0.32"
streamlit = "1.28.2"
streamlit-chat = "0.1.1"
langchain-openai = "^0.1.3"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"